requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
celery~=5.5.3
typing_extensions~=4.15.0
pydub~=0.25.1
beautifulsoup4~=4.13.5
soupsieve~=2.8
xlrd~=2.0.2
openpyxl~=3.1.5
redis~=5.2.1
amqp~=5.3.1
vine~=5.1.0
kombu~=5.5.4
billiard~=4.2.1
packaging~=25.0
click~=8.2.1
python-dateutil~=2.9.0.post0
pydantic~=2.11.7
click-repl~=0.3.0
click-plugins~=1.1.1.2
wcwidth~=0.2.13
six~=1.17.0
et_xmlfile~=2.0.0
prompt_toolkit~=3.0.52
textblob==0.18.0.post0
stripe==12.5.1
elevenlabs==2.17.0
gunicorn==23.0.0
